refactor: move class C from package p5 to package p6
