+++ src/p5/C.java
@@ -1,11 +1,1 @@
 package p5;
-
-public class C {
-  public void m3() {
-    System.out.println("m3");
-  }
-
-  public void m4() {
-    System.out.println("m4");
-  }
-}
