+++ src/p6/C.java
@@ -1,0 +1,11 @@
+package p6;
+
+public class C {
+	public void m3() {
+		System.out.println("m3");
+	}
+
+	public void m4() {
+		System.out.println("m4");
+	}
+}
